Reorganize birthday command file structure
