+++ src/commands/birthday/subcommands/index.ts
@@ -1,0 +1,4 @@
+export { default as list } from './_list';
+export { default as set } from './_set';
+export { default as show } from './_show';
+export { default as unset } from './_unset';
